Remove high level commander

See merge request polytechnique-montr-al/inf3995/20221/equipe-106/INF3995-106!138

+++ embedded/simulation/src/simulation_sensors.cc
@@ -50,6 +50,7 @@ SimulationSensors::SimulationSensors(CCrazyflieSensing* ccrazyflieSensing)
 [[nodiscard]] float SimulationSensors::getBatteryLevel(
     bool /*isInMission*/) const {
   constexpr float kToPercent = 100;
-  return m_ccrazyflieSensing->m_pcBattery->GetReading().AvailableCharge *
+  return static_cast<float>(
+             m_ccrazyflieSensing->m_pcBattery->GetReading().AvailableCharge) *
          kToPercent;
 }
